Fixed the issue: Ctrl C (killing the terminal) was not being executed.
Modified the statement from listener.join() to listener.start() as the former was 'blocking' the code (not letting the rest of the code being run.
Start() makes it a non-blocking code, making the listening a background process (thread) while the rest of the code runs.

To fix: Escape key is not executed though defined. Need to get more clarity.