Fix issue: Callback not having memory of the the angle difference becoming < 0.1.
There was a brief moment where the velocity is set to zero. But as the callback is continous, every time a pose message comes in, it starts afresh without remembering that the turtle aligned for that instant.

Implement a state flag so that this memory is saved.

Need to check for linear motion condtion.