Simplify logic
Previously only the last pressed key was disabled while still allowing the robot to move freely. But need to re-evaulate that.
turtle is now forced to turn back in.

Need to work on previous logic where last pressed key was checked.
Too many if conditions. Need to fix that. For now turtle doesnt cross limits.